Fix Design code editor theme colors
