+++ templates/design/app/components/design/CodeWorkbenchHost.test.ts
@@ -2,6 +2,8 @@ import { readFileSync } from "node:fs";
 
 import { describe, expect, it } from "vitest";
 
+import { normalizeMonacoThemeColor } from "./code-workbench-theme";
+
 describe("CodeWorkbenchHost theming", () => {
   it("passes native theme tokens into the Monaco workbench", () => {
     const source = readFileSync(
@@ -12,6 +14,7 @@ describe("CodeWorkbenchHost theming", () => {
     expect(source).toContain("readCodeWorkbenchTheme");
     expect(source).toContain("monaco-editor");
     expect(source).toContain("defineMonacoTheme");
+    expect(source).toContain("normalizeMonacoThemeColor");
     expect(source).toContain("--workbench-bg");
     expect(source).toContain("theme,");
     expect(source).not.toContain("<textarea");
@@ -19,6 +22,18 @@ describe("CodeWorkbenchHost theming", () => {
     expect(source).not.toContain("#0f1115");
     expect(source).not.toContain("#0b0d12");
     expect(source).not.toContain("bg-[#0b0d12]");
+  });
+
+  it("normalizes computed CSS colors before passing them to Monaco", () => {
+    expect(normalizeMonacoThemeColor("rgb(230, 230, 230)")).toBe("#e6e6e6");
+    expect(normalizeMonacoThemeColor("rgba(14, 165, 233, 0.4)")).toBe(
+      "#0ea5e966",
+    );
+    expect(normalizeMonacoThemeColor("rgb(90% 90% 90% / 50%)")).toBe(
+      "#e6e6e680",
+    );
+    expect(normalizeMonacoThemeColor("#fff")).toBe("#ffffff");
+    expect(normalizeMonacoThemeColor("var(--workbench-fg)")).toBeUndefined();
   });
 
   it("keeps workbench file selection, preview, and draft base versions independent", () => {
