+++ templates/design/changelog/2026-07-02-fixed-a-crash-when-opening-the-code-editor.md
@@ -1,0 +1,6 @@
+---
+type: fixed
+date: 2026-07-02
+---
+
+Fixed a crash when opening the code editor.
